src/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
Flask==3.0.0
